fix: return less explicit message on login error
Previously, the message returned on a login error explained if the issue
was with the username not existing, or if the password didn't match the
username. The new message regroups both cases for added security.

+++ src/controllers/user.php
@@ -16,23 +16,16 @@ class UserController extends Controller
 				$password = $_POST['password'];
 				$user = UserQuery::create()->findOneByLogin($login);
 				$passwordHashed = crypt($password, '$5$rounds=5000$'.md5($password).'$');
-				if ($user != null)
+				if ($user != null && $passwordHashed == $user->getPassword())
 				{
-					if ($passwordHashed == $user->getPassword())
-					{
-						SessionUtils::destroy();
-						session_start();
-						SessionUtils::set('user-login', $user->getLogin());
-						$this->redirect('');
-					}
-					else
-					{
-						$errors[] = 'Wrong password for this username.';
-					}
+					SessionUtils::destroy();
+					session_start();
+					SessionUtils::set('user-login', $user->getLogin());
+					$this->redirect('');
 				}
 				else
 				{
-					$errors[] = 'Wrong username';
+					$errors[] = 'Login failed. Invalid username or password.';
 				}
 			}
 			$template = $this->loadView('user_login_view');
